+++ include/application.hpp
@@ -5,6 +5,7 @@
 #include <memory>
 
 #include "i_form.hpp"
+#include "i_object.hpp"
 #include "object_type.hpp"
 
 // <vcc:customHeader sync="RESERVE" gen="RESERVE">
@@ -15,7 +16,7 @@ using namespace vcc;
 class Application
 {
     private:
-        std::set<std::shared_ptr<IForm>> _Forms;
+        std::set<std::shared_ptr<IObject>> _Forms;
 
     public:
         Application() = default;
@@ -23,11 +24,12 @@ class Application
  
         static void Run();
 
-        static std::shared_ptr<IForm> CreateForm(const ObjectType &objectType);
-        static std::shared_ptr<IForm> GetFormSharedPtr(const IForm *form);
-        static bool IsFormPresent(const IForm *form);
-        static bool IsFormClosable(const IForm *form);
-        static bool CloseForm(const IForm *form, const bool &isForce = false);
+        static std::shared_ptr<IObject> CreateForm(const ObjectType &objectType);
+        static std::shared_ptr<IObject> GetFormSharedPtr(IObject *IObject);
+        static const IForm *GetIFormPtrFromIObject(IObject *obj);
+        static bool IsFormPresent(IObject *form);
+        static bool IsFormClosable(IObject *form);
+        static bool CloseForm(IObject *form, const bool &isForce = false);
 
         // <vcc:custom sync="RESERVE" gen="RESERVE">
         // </vcc:custom>