+++ include/manager/vpg_cpp_generation_manager.hpp
@@ -8,11 +8,16 @@
 
 using namespace vcc;
 
-class VPGCppGenerationManager : public VPGBaseGenerationManager<VPGCppGenerationManager>
+class VPGCppGenerationManager : public VPGBaseGenerationManager
 {
     public:
         VPGCppGenerationManager(std::shared_ptr<LogConfig> logConfig, std::wstring workspace, std::shared_ptr<VPGGenerationOption> option) : VPGBaseGenerationManager(logConfig, workspace, option) {}
         virtual ~VPGCppGenerationManager() {}
+
+        virtual std::shared_ptr<IObject> Clone() const override
+        {
+            return std::make_shared<VPGCppGenerationManager>(*this);
+        }
 
         virtual void Add() const override;
 };